# demos/01-dictionary/backend/shipping-service/docs/swagger.yaml
basePath: /
definitions:
...
  rest.errorResponse:
    properties:
      error:
        type: string
    type: object
  rest.jsEvent:
    properties:
      payload:
        type: object
      seq:
        type: integer
      subject:
        type: string
      timestamp:
        type: string
    type: object
...
host: localhost:7200
info:
  contact: {}
  description: Shipping domain backend for the NATS Tech Lab POC. Demonstrates JetStream
    event sourcing, NATS KV projections (Shape A / Shape B), and pure event reconstruction
    (Shape C).
  title: EventSourcing CQRS POC — Shipping API
  version: "1.0"
paths:
...
  /api/jetstream/replay:
    get:
      description: 'Returns every currently-retained raw JetStream message from one
        account''s stream as a single JSON array, snapshotted at request time. Replaces
        replayJetStream''s SSE full-history half — the Admin UI now does one bootstrap
        fetch here, then subscribes to notify.{context}.shipping.raw.{entity}.{event}
        (eventhandler.publishRawNotify) for anything published afterward, instead
        of holding an SSE connection open. Because this snapshot is backend-mediated
        it works for ANY account listStreams reports, not just the browser''s own
        — but the notify.* live tail does not: NATS enforces account isolation at
        the server, so a browser can only subscribe within the account its own connection
        authenticated as.'
      parameters:
      - description: NATS account (a known tenant name, or \
        in: query
        name: account
        type: string
      - description: Stream name (default SHIPPING)
        in: query
        name: stream
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rest.jsEvent'
            type: array
        "400":
          description: Unknown account or stream
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: JetStream one-shot replay (Phase 23)
      tags:
      - streams
...
swagger: "2.0"